Factored out availableOptions / blueprints from commands.
