+++ packages/angular-cli/commands/version.options.ts
@@ -1,0 +1,3 @@
+export const availableOptions: any = [
+  { name: 'verbose', type: Boolean, 'default': false }
+];
